Create Menu Arrays Done
ageCheck() is broken, I am going to make an ageCheck() branch and try a few different ways. I also want to clean up the menu methods. Ideally I will have the arrays set up in their own island and then one method with the loop but for the moment it works